Scala JUnit launcher with expression evaluator
Scala JUnit launcher is working pretty cool but unfortunately
there is no possibility to evaluate expressions when stopped
in breakpoint. The scala expression evaluator needs in its
classpath additional classes which are not exposed by now.
This commit makes Scala JUnit launcher available for Scala
expression evaluator.
