+++ org.scala-ide.sdt.debug.tests/META-INF/MANIFEST.MF
@@ -8,6 +8,10 @@ Fragment-Host: org.scala-ide.sdt.debug
 Bundle-RequiredExecutionEnvironment: JavaSE-1.6
 Require-Bundle: org.scala-lang.scala-library,
  org.eclipse.equinox.weaving.aspectj,
+ org.eclipse.pde.launching;bundle-version="[3.6.0,4.0.0)",
+ org.eclipse.jdt.junit4.runtime,
+ org.eclipse.jdt.junit.core,
+ org.eclipse.jdt.junit.runtime,
  org.scala-ide.sdt.core,
  org.junit,
  org.eclipse.jdt.junit4.runtime,
