+++ org.scala-ide.sdt.debug.tests/src/org/scalaide/debug/internal/ScalaDebugTestSuite.scala
@@ -9,6 +9,8 @@ import org.scalaide.debug.internal.editor.StackFrameVariableOfTreeFinderTest
 import org.scalaide.debug.internal.hcr.HotCodeReplaceTest
 import org.scalaide.debug.internal.launching.LibraryJarInBootstrapTest
 import org.scalaide.debug.internal.launching.RemoteConnectorTest
+import org.scalaide.debug.internal.launching.ScalaApplicationLaunchConfigurationDelegateTest
+import org.scalaide.debug.internal.launching.ScalaJUnitLaunchConfigurationDelegateTest
 import org.scalaide.debug.internal.model.DebugTargetTerminationTest
 import org.scalaide.debug.internal.model.MethodClassifierUnitTest
 import org.scalaide.debug.internal.model.ScalaDebugCacheTest
@@ -17,7 +19,6 @@ import org.scalaide.debug.internal.model.ScalaDebugTargetTest
 import org.scalaide.debug.internal.model.ScalaStackFrameTest
 import org.scalaide.debug.internal.model.ScalaThreadTest
 import org.scalaide.debug.internal.model.ScalaValueTest
-import org.scalaide.debug.internal.launching.ScalaApplicationLaunchConfigurationDelegateTest
 
 /**
  * Junit test suite for the Scala debugger.
@@ -42,6 +43,7 @@ import org.scalaide.debug.internal.launching.ScalaApplicationLaunchConfiguration
     classOf[ScalaDebugCacheTest],
     classOf[StackFrameVariableOfTreeFinderTest],
     classOf[HotCodeReplaceTest],
-    classOf[ScalaApplicationLaunchConfigurationDelegateTest]
+    classOf[ScalaApplicationLaunchConfigurationDelegateTest],
+    classOf[ScalaJUnitLaunchConfigurationDelegateTest]
     ))
 class ScalaDebugTestSuite
