+++ org.scala-ide.sdt.debug.tests/src/org/scalaide/debug/internal/launching/ScalaJUnitLaunchConfigurationDelegateTest.scala
@@ -3,18 +3,20 @@ package org.scalaide.debug.internal.launching
 import org.eclipse.debug.core.ILaunchManager
 import org.junit.After
 import org.junit.Before
+import org.junit.Ignore
 import org.junit.Test
 import org.scalaide.core.testsetup.IProjectHelpers
+import org.scalaide.core.testsetup.SdtTestConstants
 import org.scalaide.core.testsetup.TestProjectSetup
 
-import ScalaApplicationLaunchConfigurationDelegateTest.file
-import ScalaApplicationLaunchConfigurationDelegateTest.project
+import ScalaJUnitLaunchConfigurationDelegateTest.file
+import ScalaJUnitLaunchConfigurationDelegateTest.project
 
-object ScalaApplicationLaunchConfigurationDelegateTest
-  extends TestProjectSetup("launchDelegate", bundleName = "org.scala-ide.sdt.debug.tests")
+object ScalaJUnitLaunchConfigurationDelegateTest
+  extends TestProjectSetup("launchJUnitDelegate", bundleName = "org.scala-ide.sdt.debug.tests")
 
-class ScalaApplicationLaunchConfigurationDelegateTest extends LaunchUtils with IProjectHelpers {
-  import ScalaApplicationLaunchConfigurationDelegateTest._
+class ScalaJUnitLaunchConfigurationDelegateTest extends LaunchUtils with IProjectHelpers {
+  import ScalaJUnitLaunchConfigurationDelegateTest._
   override val launchConfigurationName = "launchDelegate"
   private val fileWithLaunchEffectName = "launchDelegate.result"
 
@@ -26,6 +28,7 @@ class ScalaApplicationLaunchConfigurationDelegateTest extends LaunchUtils with I
     file(fileWithLaunchEffectName).delete( /*force = */ true, /*monitor = */ null)
   }
 
+  @Ignore(SdtTestConstants.TestRequiresGuiSupport)
   @Test def shouldLaunchTestInDebugMode(): Unit = {
     whenApplicationWasLaunchedFor(project, ILaunchManager.DEBUG_MODE) {
       assertLaunchEffect(project, ILaunchManager.DEBUG_MODE, file(fileWithLaunchEffectName))
