+++ org.scala-ide.sdt.debug.tests/test-workspace/launchDelegate/launchDelegate.launch
@@ -8,7 +8,6 @@
 </listAttribute>
 <mapAttribute key="org.eclipse.debug.core.preferred_launchers">
 <mapEntry key="[debug]" value="scala.application.new"/>
-<mapEntry key="[run]" value="scala.application.new"/>
 </mapAttribute>
 <stringAttribute key="org.eclipse.jdt.launching.MAIN_TYPE" value="test.ScalaTest"/>
 <stringAttribute key="org.eclipse.jdt.launching.PROJECT_ATTR" value="launchDelegate"/>
