+++ org.scala-ide.sdt.debug.tests/test-workspace/launchJUnitDelegate/launchDelegate.launch
@@ -1,0 +1,19 @@
+<?xml version="1.0" encoding="UTF-8" standalone="no"?>
+<launchConfiguration type="org.eclipse.jdt.junit.launchconfig">
+<listAttribute key="org.eclipse.debug.core.MAPPED_RESOURCE_PATHS">
+<listEntry value="/launchJUnitDelegate/src/test/ScalaTest.scala"/>
+</listAttribute>
+<listAttribute key="org.eclipse.debug.core.MAPPED_RESOURCE_TYPES">
+<listEntry value="1"/>
+</listAttribute>
+<mapAttribute key="org.eclipse.debug.core.preferred_launchers">
+<mapEntry key="[debug]" value="scala.junit.new"/>
+</mapAttribute>
+<stringAttribute key="org.eclipse.jdt.junit.CONTAINER" value=""/>
+<booleanAttribute key="org.eclipse.jdt.junit.KEEPRUNNING_ATTR" value="false"/>
+<stringAttribute key="org.eclipse.jdt.junit.TESTNAME" value=""/>
+<stringAttribute key="org.eclipse.jdt.junit.TEST_KIND" value="org.scala-ide.sdt.core.junit"/>
+<stringAttribute key="org.eclipse.jdt.launching.MAIN_TYPE" value="test.ScalaTest"/>
+<stringAttribute key="org.eclipse.jdt.launching.PROJECT_ATTR" value="launchJUnitDelegate"/>
+<stringAttribute key="pde.version" value="3.3"/>
+</launchConfiguration>
