+++ org.scala-ide.sdt.debug/plugin.xml
@@ -28,7 +28,7 @@
             delegate="org.scalaide.debug.internal.launching.ScalaApplicationLaunchConfigurationDelegate"
             delegateDescription="The Scala JVM Launcher supports debugging of local Scala using the new Scala debugger"
             id="scala.application.new"
-            modes="debug, run"
+            modes="debug"
             name="Scala Application (new debugger)"
             sourceLocatorId="org.eclipse.jdt.launching.sourceLocator.JavaSourceLookupDirector"
             sourcePathComputerId="org.eclipse.jdt.launching.sourceLookup.javaSourcePathComputer"
@@ -53,6 +53,16 @@
             sourceLocatorId="org.eclipse.pde.ui.launcher.PDESourceLookupDirector"
             sourcePathComputerId="org.eclipse.jdt.launching.sourceLookup.javaSourcePathComputer"
             type="org.scala-ide.ew.launcher.JunitLaunchConfig">
+      </launchDelegate>
+      <launchDelegate
+            delegate="org.scalaide.debug.internal.launching.ScalaJUnitLaunchConfigurationDelegate"
+            delegateDescription="The Scala JVM Launcher supports debugging of JUnit Scala using the new Scala debugger"
+            id="scala.junit.new"
+            modes="debug"
+            name="Scala JUnit (new debugger)"
+            sourceLocatorId="org.eclipse.jdt.launching.sourceLocator.JavaSourceLookupDirector"
+            sourcePathComputerId="org.eclipse.jdt.launching.sourceLookup.javaSourcePathComputer"
+            type="org.eclipse.jdt.junit.launchconfig">
       </launchDelegate>
     </extension>
    <extension
