+++ org.scala-ide.sdt.debug/src/org/scalaide/debug/internal/launching/ScalaDebuggerForLaunchDelegate.scala
@@ -1,0 +1,13 @@
+package org.scalaide.debug.internal.launching
+
+import org.eclipse.debug.core.ILaunchConfiguration
+import org.eclipse.jdt.launching.AbstractJavaLaunchConfigurationDelegate
+import org.eclipse.jdt.launching.IVMRunner
+import org.eclipse.pde.internal.launching.launcher.VMHelper
+
+trait ScalaDebuggerForLaunchDelegate extends AbstractJavaLaunchConfigurationDelegate {
+  override def getVMRunner(configuration: ILaunchConfiguration, mode: String): IVMRunner = {
+    val vm = VMHelper.createLauncher(configuration)
+    new StandardVMScalaDebugger(vm)
+  }
+}
